requirements.txt
alembic==1.12.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.2
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
WTForms==3.0.1
